+++ resources/views/layouts/UserAccount/admindetailUsers.blade.php
@@ -106,40 +106,66 @@
                                     {{session('message')}}
                                 @endif
                                 <!-- Single Tab Content Start -->
-                                    <div class="tab-pane fade show {{ (Route::currentRouteName() == 'adminusers')}}" id="users" role="tabpanel">
+                                    <div class="tab-pane fade show {{ (Route::currentRouteName() == 'adminusers') ? 'active' : '' }} active" id="users" role="tabpanel">
                                         <div class="myaccount-content">
                                             <h5>Utenti</h5>
+                                            <div class="blog-left-title">
+                                                <h3>Search</h3>
+                                            </div>
+                                            <div class="side-form">
+                                                <form action="{{ route('searchusersrouteAdminPanel') }}">
+                                                    <input type="text" name="search" placeholder="Cerca un articolo..." />
+                                                    <a  href="javascript:;" onclick="parentNode.submit();"><i class="fa fa-search"></i></a>
+                                                </form>
+                                            </div>
+                                            <div class="mt-3"></div>
+                                            @if($users->count() > 0)
                                             <div class="myaccount-table table-responsive text-center">
-                                                <table class="table table-bordered">
-                                                    <thead class="thead-light">
-                                                    <tr>
-                                                        <th>Nickname</th>
-                                                        <th>Phone</th>
-                                                        <th>Email</th>
-                                                        <th>Tipologia</th>
-                                                        <th>Elimina</th>
-                                                    </tr>
-                                                    </thead>
-                                                    <tbody>
-                                                    @foreach($users as $user)
+                                                    <table class="table table-bordered mt-2">
+                                                        <thead class="thead-light">
                                                         <tr>
-                                                            <td>{{$user->username}}</td>
-                                                            <td>{{$user->phone_number}}</td>
-                                                            <td>{{$user->email}}</td>
-                                                            @if($user->hasGroup('il gruppo degli utenti'))
-                                                                <td>Utente</td>
-                                                            @else
-                                                                <td>Venditore</td>
-                                                            @endif
-                                                            <td><a class="btn btn-danger" onclick="return deleteUser();"  href="{{route('user-delete', $user->id)}}"><i class="fa fa-trash"></i></a></td>
+                                                            <th>Nickname</th>
+                                                            <th>Phone</th>
+                                                            <th>Email</th>
+                                                            <th>Tipologia</th>
+                                                            <th>Elimina</th>
                                                         </tr>
+                                                        </thead>
+                                                        <tbody>
+                                                        @foreach($users as $user)
+                                                            @if(!(\App\Http\Controllers\GroupController::isAdmin($user->id)))
+                                                                <tr>
+                                                                    <td>{{$user->username}}</td>
+                                                                    <td>{{$user->phone_number}}</td>
+                                                                    <td>{{$user->email}}</td>
+                                                                    @if(!(\App\Http\Controllers\GroupController::isVendor($user->id)))
+                                                                        <td>Utente</td>
+                                                                    @else
+                                                                        <td>Venditore</td>
+                                                                    @endif
+                                                                    <td><a class="btn btn-danger" onclick="return deleteUser();"  href="{{route('user-delete', $user->id)}}"><i class="fa fa-trash"></i></a></td>
+                                                                </tr>
+                                                            @endif
+                                                        @endforeach
+                                                        {{ $users->links() }}
+                                                        </tbody>
+                                                    </table>
+                                                    @foreach($users as $user)
                                                     @endforeach
                                                     {{ $users->links() }}
-                                                    </tbody>
-                                                </table>
-                                                @foreach($users as $user)
-                                                @endforeach
-                                                {{ $users->links() }}
+                                            </div>
+                                            @else
+                                                <h3>nessun risultato di ricerca</h3>
+                                            @endif
+                                        </div>
+                                        <div class="mt-3"></div>
+                                        <div class="row">
+                                            <div class="col-lg-10"></div>
+                                            <div style="margin-right: 2.75%"></div>
+                                            <div class="col-lg-1">
+                                                <div class="buttons-back">
+                                                    <a href="{{url('/adminArea/users')}}">Indietro</a>
+                                                </div>
                                             </div>
                                         </div>
                                     </div>
@@ -227,7 +253,7 @@
                                         </div>
                                     </div>
 
-                                    <div class="tab-pane fade show {{ (Route::currentRouteName() == 'adminarticles') }} active" id="articles" role="tabpanel">
+                                    <div class="tab-pane fade show {{ (Route::currentRouteName() == 'adminarticles') }}" id="articles" role="tabpanel">
                                         <div class="myaccount-content">
                                             <h5>Articoli</h5>
                                             <div class="blog-left-title">
